requirements.txt
alabaster==0.7.12
arrow==1.1.1
Babel==2.9.1
beautifulsoup4==4.10.0
bibtexparser==1.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
docutils==0.16
fake-useragent==0.1.11
free-proxy==1.0.3
future==0.18.2
idna==3.2
imagesize==1.2.0
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
packaging==21.0
Pygments==2.10.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
scholarly==1.3.1
selenium==3.141.0
six==1.16.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==4.1.2
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stem==1.8.0
typing-extensions==3.10.0.2
urllib3==1.26.6
